+++ minn-admin.php
@@ -89,6 +89,7 @@ require_once MINN_ADMIN_DIR . 'includes/adapters/spam.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/site-status.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/user-switching.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/one-time-login.php';
+require_once MINN_ADMIN_DIR . 'includes/adapters/public-post-preview.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/regenerate-thumbnails.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/wcpdf.php';
 require_once MINN_ADMIN_DIR . 'includes/adapters/licenses.php';